requirements.txt
Pillow
requests
uvicorn
fastapi
diffusers
huggingface_hub
transformers
sentencepiece
protobuf
aiohttp
accelerate
safetensors
gradio_client
python-dotenv
boto3